abstracted searching profile by user id into a class  method for Profile model. Also made a small cleanup of Profile controller to use this function. Went into user model and separated the old method for deleting gallery and applying it to profile instance
