+++ server/models/Profile.js
@@ -1,10 +1,12 @@
 const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
+const { deleteFile } = require("../utils/aws")
+
 
 const profileSchema = new Schema({
-  userId: {
+  user: {
     type: Schema.Types.ObjectId,
-    ref: "User",
+    ref: "user",
   },
   firstName: {
     type: String,
@@ -52,6 +54,43 @@ const profileSchema = new Schema({
   },
   gallery: [String],
 });
+
+profileSchema.statics.findByUserIdPopulated = async function (userId) {
+  return (
+    await this.findOne({ userId: userId }).populate("user", "_id username avatar")
+  )
+}
+
+profileSchema.methods.addToGallery = async function (newImages) {
+  this.gallery.push(...newImages)
+  await this.save()
+}
+
+profileSchema.methods.deleteFromGallery = async function (links) {
+  let tempGallery = new Set(this.gallery)
+  let filesToBeDeleted = []
+
+
+  links.forEach(element => {
+    if (tempGallery.delete(element)) {
+      filesToBeDeleted.push(element)
+    }
+  })
+
+
+  this.gallery = [...tempGallery]
+
+  await this.save()
+
+  if (filesToBeDeleted.length) {
+    await cleanUpAWSFolder([...filesToBeDeleted])
+  }
+
+}
+
+const cleanUpAWSFolder = (keys) => {
+  deleteFile(keys.map(element => element.split(".s3.amazonaws.com/").pop()))
+}
 
 
 module.exports = Profile = mongoose.model("profile", profileSchema);
