+++ server/models/User.js
@@ -48,33 +48,6 @@ userSchema.methods.editAvatar = async function (url) {
 }
 
 
-userSchema.methods.addToGallery = async function (newImages) {
-  this.gallery.push(...newImages)
-  await this.save()
-
-}
-
-userSchema.methods.deleteFromGallery = async function (links) {
-  let tempGallery = new Set(this.gallery)
-  let filesToBeDeleted = []
-
-
-  links.forEach(element => {
-    if (tempGallery.delete(element)) {
-      filesToBeDeleted.push(element)
-    }
-  })
-
-
-  this.gallery = [...tempGallery]
-
-  await this.save()
-
-  if (filesToBeDeleted.length) {
-    await cleanUpAWSFolder([...filesToBeDeleted])
-  }
-
-}
 
 userSchema.methods.deleteAvatar = async function () {
   const deletedImage = this.avatar
@@ -84,13 +57,7 @@ userSchema.methods.deleteAvatar = async function () {
 }
 
 const cleanUpAWSFolder = (keys) => {
-  if (Array.isArray(keys)) {
-    deleteFile(keys.map(element => element.split(".s3.amazonaws.com/").pop()))
-  }
-  else {
-    deleteFile(keys.split(".s3.amazonaws.com/")[1])
-
-  }
+  deleteFile(keys.split(".s3.amazonaws.com/")[1])
 }
 
 module.exports = User = mongoose.model("user", userSchema);
