+++ server/serializers/userSerializer.js
@@ -4,7 +4,6 @@ function userSerializer(user){
         username: user.username,
         email: user.email,
         avatar: user.avatar,
-        gallery: user.gallery
     }
     )
 }
